Release v0.2.1-alpha
Release v0.2.1-alpha

* Bitcoin deploy notes (#319)

* Boot repackage was disabled (#315)

* add more logs to eth registration

* add more logs to eth registration

* Bitcoin rollback (#321)

* Bitcoin rollback in case of network failures (#323)

* Bitcoin multisig address safe time source (#325)

* Ignoring failed Iroha account creation command (#324)

* Kill deposit if Iroha is not functioning (#326)

* add automatic relay deployment logic

* Separate shadow jar (#327)

* optimized import

* fix config helper

* add Sora iroha account to genesis and tests for Sora integration

* fix dogecoin precision (actual 9)

* rename getAccountAsset -> getAccountAssetBalance

* change getAccountAssets -> getIrohaAccountBalance

* Bitcoin dynamic fee rate (#329)

* Feature/healthcheck (#333)

* Add readme

* Add healthcheck endpoints

* Fix test compiler error

* Port increment

* Remove registration endpoint

* Transition to IPJ, Iroha rc2 version update (#335)

* Feature/iroha pure java (#307)

* move crypto to Iroha Pure Java library

* Transition to IPJ, Iroha rc2 version update (#330)

* Transition to IPJ

* Iroha rc1 update

* Test fixes

* Iroha rc2 update

* Updated permissions, roles, testing account

* Test fixes

* Update batches

* Fixed status streaming and batch testing

* Switched to ordered batches and to Query API of IPJ

* More queries refactoring, changed wrapper

* Unit tests fixes

* Ignored one more unit test

* Integration tests fixes

* fixup! Integration tests fixes

* fixup! Integration tests fixes

* fixup! fixup! Integration tests fixes

* Config test repaired

* Environment variables configs (#336)

* Added prioritized environment variables config loading

* Fixed working directory of eth services (#338)

* Fixed notary pub key (#340)

* Fixed notary pub key

* Deleted unused notary keys

* fix 'mkdir -p' in Jenkins build-logs (#341)

* Bitcoin services stabilization after migration to IPJ (#342)

* Automatic BTC addresses generation (#337)

* Bitcoin-cli in gradle (#344)

* Feature/tester permissions (#343)

* add permissions for tester account

* fix duplication of can_read_assets

* Fixed PeerGroup.stop()

* Fixing IPJ listener bug (#346)

* Relevant docker files (#345)

* Updated docker files, added bitcoin services

* Deleted variables

* Fixed jenkinsfile variables names

* Fixed main classes

* Fix wait for iroha bug (#348)

* Fix wait for iroha bug

* Add network to dev

* Fix typo

* Fix btc docker. Move dw and address geneartion from single to full (#349)

* Email notifications

* Added build commands to Jenkinsfile (#351)

* Feature/registration (#334)

* Fixed message

* Fixing build (#352)

* Fixed peer group close() (#354)

* SMTP dumbster integration test (#355)

+++ eth/src/main/kotlin/provider/eth/EthTokensProvider.kt
@@ -9,7 +9,7 @@ interface EthTokensProvider {
     fun getTokens(): Result<Map<String, String>, Exception>
 
     /** Return token precision by asset name */
-    fun getTokenPrecision(name: String): Result<Short, Exception>
+    fun getTokenPrecision(name: String): Result<Int, Exception>
 
     /** Return token precision by asset name */
     fun getTokenAddress(name: String): Result<String, Exception>
